+++ src/index.tsx
@@ -1,7 +1,10 @@
 import React from "react";
+
 import ReactDOM from "react-dom/client";
-import { Layout } from "~/components/common/Layout";
+
+import { Layout } from "~/components/common/layout/Layout";
 import { Route } from "~/components/common/route/Route";
+
 import "~/index.modules.scss";
 
 const root = ReactDOM.createRoot(
